Merge remote-tracking branch 'origin/master' into BambaArchitectureRefactor
This commit is TOTALLY BROKEN! It's got residual conflict markers in some
files (llama-kv-cache.h) and I just totally blew away all changes in
llama.cpp and didn't attempt to resurrect them.

It's become clear that the refactors I'm trying to incorporate into this
branch are themselves trying to solve the hybrid cache problem, but doing
so in a completely different way from my approach of maintaining
side-by-side instances of the common cache structure. I'm going to pivot to
try to get things working better with the current refactor from scratch.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>